requirements.txt
alembic==1.11.1
arviz==0.15.1
cachetools==5.3.1
calibr8==7.1.0
chaospy==4.3.13
cloudpickle==2.2.1
cmaes==0.9.1
colorlog==6.7.0
cons==0.4.6
contourpy==1.1.0
cycler==0.11.0
dill==0.3.6
emcee==3.1.4
etuples==0.3.9
fastmat==0.2.2
fastprogress==1.0.3
filelock==3.12.2
fonttools==4.40.0
greenlet==2.0.2
h5netcdf==1.2.0
h5py==3.9.0
importlib-metadata==6.8.0
importlib-resources==6.0.0
joblib==1.3.1
julia==0.6.1
kaleido==0.2.1
kiwisolver==1.4.4
logical-unification==0.4.6
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.7.2
miniKanren==1.0.3
multipledispatch==1.0.0
multiprocess==0.70.14
murefi==5.2.0
numpoly==1.2.11
numpy==1.25.1
optuna==3.2.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pip==23.1.2
plotly==5.15.0
pyDOE2==1.3.0
pygpc==0.3.3
pymc==5.6.0
pyparsing==3.0.9
pytensor==2.12.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
SALib==1.4.7
scikit-learn==1.3.0
scipy==1.11.1
setuptools==67.8.0
six==1.16.0
smt==2.0
spotpy==1.6.2
SQLAlchemy==2.0.18
tenacity==8.2.2
threadpoolctl==3.1.0
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
wheel==0.38.4
xarray==2023.6.0
xarray-einstats==0.6.0
zipp==3.16.0
